SSM-2692: extend the timestamp interval to use a long integer.  Limiting user input to <100 years.
git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@26826 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ modules/identity/src/main/java/com/l7tech/identity/ldap/LdapIdentityProviderConfig.java
@@ -445,7 +445,7 @@ public class LdapIdentityProviderConfig extends IdentityProviderConfig implement
      * @return The maximum age or null
      */
     @Transient
-    public Integer getGroupCacheMaxAge() {
+    public Long getGroupCacheMaxAge() {
         return getProperty(GROUP_CACHE_MAX_AGE);
     }
 
@@ -454,7 +454,7 @@ public class LdapIdentityProviderConfig extends IdentityProviderConfig implement
      *
      * @param maxAge The maximum age (null for default, 0 for no caching)
      */
-    public void setGroupCacheMaxAge( final Integer maxAge ) {
+    public void setGroupCacheMaxAge( final Long maxAge ) {
         setProperty(GROUP_CACHE_MAX_AGE, maxAge);
     }
